Multiplayer bug fix
Fixed bug where players would be moved to the next player slot (PLayer 1 to player 2) after failing and retrying the level. Side note, unable to fix controls issue without heavy changes to how it is designed, so for now the players must press enter on a keyboard once they are on the main menu after a game over, unable to use controller to navigate it.